sensors_py: Move simple functions to eigen_pybind
