+++ bindings/pydrake/systems/BUILD.bazel
@@ -110,7 +110,9 @@ drake_pybind_library(
 drake_pybind_library(
     name = "sensors_py",
     cc_deps = [
+        "//bindings/pydrake/systems:systems_pybind",
         "//bindings/pydrake/util:cpp_template_pybind",
+        "//bindings/pydrake/util:eigen_pybind",
     ],
     cc_so_name = "sensors",
     cc_srcs = ["sensors_py.cc"],
