+++ bindings/pydrake/util/eigen_pybind.h
@@ -3,6 +3,8 @@
 #include <Eigen/Dense>
 #include "pybind11/eigen.h"
 
+#include "drake/common/eigen_types.h"
+
 namespace drake {
 namespace pydrake {
 
@@ -13,6 +15,26 @@ namespace pydrake {
 template <typename Derived>
 auto ToEigenRef(Eigen::VectorBlock<Derived>* derived) {
   return Eigen::Ref<Derived>(*derived);
+}
+
+/// Converts a raw array to a numpy array.
+template <typename T>
+py::object ToArray(T* ptr, int size, py::tuple shape) {
+  // Create flat array to be reshaped in numpy.
+  using Vector = VectorX<T>;
+  Eigen::Map<Vector> data(ptr, size);
+  return py::cast(
+      Eigen::Ref<Vector>(data), py_reference).attr("reshape")(shape);
+}
+
+/// Converts a raw array to a numpy array (`const` variant).
+template <typename T>
+py::object ToArray(const T* ptr, int size, py::tuple shape) {
+  // Create flat array to be reshaped in numpy.
+  using Vector = const VectorX<T>;
+  Eigen::Map<Vector> data(ptr, size);
+  return py::cast(
+      Eigen::Ref<Vector>(data), py_reference).attr("reshape")(shape);
 }
 
 }  // namespace pydrake
